+++ disassemblers/ofrak_ghidra/src/ofrak_ghidra/ghidra_scripts/CreateMemoryBlocks.java
@@ -1,5 +1,4 @@
 import com.google.common.base.Strings;
-import com.google.common.base.Strings;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import com.sun.net.httpserver.HttpServer;
@@ -66,8 +65,8 @@ public class CreateMemoryBlocks extends HeadlessScript {
 
             String[] memRegionInfo = memRegionRaw.split("!");
 
-            int address = Integer.parseInt(memRegionInfo[0]);
-            int size = Integer.parseInt(memRegionInfo[1]);
+            long address = Long.parseLong(memRegionInfo[0]);
+            long size = Long.parseLong(memRegionInfo[1]);
             String permissions = memRegionInfo[2];
             String name = memRegionInfo[3];
             int offset = Integer.parseInt(memRegionInfo[4]);